requirements.txt
selenium
pandas
numpy
requests
sqlalchemy
boto3
python-dotenv

